+++ Source/Core/GoogleEarthEnterpriseTerrainProvider.js
@@ -5,6 +5,7 @@ define([
     './defaultValue',
     './defined',
     './defineProperties',
+    './deprecationWarning',
     './DeveloperError',
     './Event',
     './GeographicTilingScheme',
@@ -15,9 +16,11 @@ define([
     './loadArrayBuffer',
     './Math',
     './Rectangle',
+    './Request',
+    './RequestState',
+    './RequestType',
     './RuntimeError',
     './TaskProcessor',
-    './throttleRequestByServer',
     './TileProviderError'
 ], function(
     when,
@@ -25,6 +28,7 @@ define([
     defaultValue,
     defined,
     defineProperties,
+    deprecationWarning,
     DeveloperError,
     Event,
     GeographicTilingScheme,
@@ -35,9 +39,11 @@ define([
     loadArrayBuffer,
     CesiumMath,
     Rectangle,
+    Request,
+    RequestState,
+    RequestType,
     RuntimeError,
     TaskProcessor,
-    throttleRequestByServer,
     TileProviderError) {
     'use strict';
 
@@ -152,6 +158,7 @@ define([
 
         this._terrainCache = new TerrainCache();
         this._terrainPromises = {};
+        this._terrainRequests = {};
 
         this._errorEvent = new Event();
 
@@ -339,9 +346,7 @@ define([
      * @param {Number} x The X coordinate of the tile for which to request geometry.
      * @param {Number} y The Y coordinate of the tile for which to request geometry.
      * @param {Number} level The level of the tile for which to request geometry.
-     * @param {Boolean} [throttleRequests=true] True if the number of simultaneous requests should be limited,
-     *                  or false if the request should be initiated regardless of the number of requests
-     *                  already in progress.
+     * @param {Request} [request] The request object. Intended for internal use only.
      * @returns {Promise.<TerrainData>|undefined} A promise for the requested geometry.  If this method
      *          returns undefined instead of a promise, it is an indication that too many requests are already
      *          pending and the request will be retried later.
@@ -349,7 +354,7 @@ define([
      * @exception {DeveloperError} This function must not be called before {@link GoogleEarthEnterpriseProvider#ready}
      *            returns true.
      */
-    GoogleEarthEnterpriseTerrainProvider.prototype.requestTileGeometry = function(x, y, level, throttleRequests) {
+    GoogleEarthEnterpriseTerrainProvider.prototype.requestTileGeometry = function(x, y, level, request) {
         //>>includeStart('debug', pragmas.debug)
         if (!this._ready) {
             throw new DeveloperError('requestTileGeometry must not be called before the terrain provider is ready.');
@@ -434,23 +439,31 @@ define([
 
         // Load that terrain
         var terrainPromises = this._terrainPromises;
+        var terrainRequests = this._terrainRequests;
         var url = buildTerrainUrl(this, q, terrainVersion);
-        var promise;
+        var sharedPromise;
+        var sharedRequest;
         if (defined(terrainPromises[q])) { // Already being loaded possibly from another child, so return existing promise
-            promise = terrainPromises[q];
+            sharedPromise = terrainPromises[q];
+            sharedRequest = terrainRequests[q];
         } else { // Create new request for terrain
-            var requestPromise;
-            throttleRequests = defaultValue(throttleRequests, true);
-            if (throttleRequests) {
-                requestPromise = throttleRequestByServer(url, loadArrayBuffer);
-                if (!defined(requestPromise)) {
-                    return undefined; // Throttled
-                }
-            } else {
-                requestPromise = loadArrayBuffer(url);
+            if (typeof request === 'boolean') {
+                deprecationWarning('throttleRequests', 'The throttleRequest parameter for requestTileGeometry was deprecated in Cesium 1.35.  It will be removed in 1.37.');
+                request = new Request({
+                    throttle : request,
+                    throttleByServer : request,
+                    type : RequestType.TERRAIN
+                });
             }
 
-            promise = requestPromise
+            sharedRequest = request;
+            var requestPromise = loadArrayBuffer(url, undefined, sharedRequest);
+
+            if (!defined(requestPromise)) {
+                return undefined; // Throttled
+            }
+
+            sharedPromise = requestPromise
                 .then(function(terrain) {
                     if (defined(terrain)) {
                         return taskProcessor.scheduleTask({
@@ -482,22 +495,20 @@ define([
                     }
 
                     return when.reject(new RuntimeError('Failed to load terrain.'));
-                })
-                .otherwise(function(error) {
-                    info.terrainState = TerrainState.NONE;
-                    return when.reject(error);
                 });
 
-            terrainPromises[q] = promise; // Store promise without delete from terrainPromises
+            terrainPromises[q] = sharedPromise; // Store promise without delete from terrainPromises
+            terrainRequests[q] = sharedRequest;
 
             // Set promise so we remove from terrainPromises just one time
-            promise = promise
+            sharedPromise = sharedPromise
                 .always(function() {
                     delete terrainPromises[q];
+                    delete terrainRequests[q];
                 });
         }
 
-        return promise
+        return sharedPromise
             .then(function() {
                 var buffer = terrainCache.get(quadKey);
                 if (defined(buffer)) {
@@ -509,10 +520,17 @@ define([
                         negativeAltitudeExponentBias: metadata.negativeAltitudeExponentBias,
                         negativeElevationThreshold: metadata.negativeAltitudeThreshold
                     });
-                } else {
-                    info.terrainState = TerrainState.NONE;
-                    return when.reject(new RuntimeError('Failed to load terrain.'));
                 }
+
+                return when.reject(new RuntimeError('Failed to load terrain.'));
+            })
+            .otherwise(function(error) {
+                if (sharedRequest.state === RequestState.CANCELLED) {
+                    request.state = sharedRequest.state;
+                    return when.reject(error);
+                }
+                info.terrainState = TerrainState.NONE;
+                return when.reject(error);
             });
     };
 
@@ -569,7 +587,12 @@ define([
 
         if (metadata.isValid(quadKey)) {
             // We will need this tile, so request metadata and return false for now
-            metadata.populateSubtree(x, y, level);
+            var request = new Request({
+                throttle : true,
+                throttleByServer : true,
+                type : RequestType.TERRAIN
+            });
+            metadata.populateSubtree(x, y, level, request);
         }
         return false;
     };