+++ Specs/Core/GoogleEarthEnterpriseMetadataSpec.js
@@ -7,6 +7,7 @@ defineSuite([
     'Core/defaultValue',
     'Core/loadWithXhr',
     'Core/Math',
+    'Core/Request',
     'ThirdParty/when'
 ], function(
     GoogleEarthEnterpriseMetadata,
@@ -16,6 +17,7 @@ defineSuite([
     defaultValue,
     loadWithXhr,
     CesiumMath,
+    Request,
     when) {
     'use strict';
 
@@ -114,7 +116,7 @@ defineSuite([
     it('populateSubtree', function() {
         var quad = '0123';
         var index = 0;
-        spyOn(GoogleEarthEnterpriseMetadata.prototype, 'getQuadTreePacket').and.callFake(function(quadKey, version) {
+        spyOn(GoogleEarthEnterpriseMetadata.prototype, 'getQuadTreePacket').and.callFake(function(quadKey, version, request) {
             quadKey = defaultValue(quadKey, '') + index.toString();
             this._tileInfo[quadKey] = new GoogleEarthEnterpriseTileInformation(0xFF, 1, 1, 1);
             index = (index + 1) % 4;
@@ -125,17 +127,20 @@ defineSuite([
         var metadata = new GoogleEarthEnterpriseMetadata({
             url : 'http://test.server'
         });
+        var request = new Request({
+            throttle : true
+        });
         return metadata.readyPromise
             .then(function() {
                 var tileXY = GoogleEarthEnterpriseMetadata.quadKeyToTileXY(quad);
-                return metadata.populateSubtree(tileXY.x, tileXY.y, tileXY.level);
+                return metadata.populateSubtree(tileXY.x, tileXY.y, tileXY.level, request);
             })
             .then(function() {
                 expect(GoogleEarthEnterpriseMetadata.prototype.getQuadTreePacket.calls.count()).toEqual(4);
-                expect(GoogleEarthEnterpriseMetadata.prototype.getQuadTreePacket).toHaveBeenCalledWith('', 1, false);
-                expect(GoogleEarthEnterpriseMetadata.prototype.getQuadTreePacket).toHaveBeenCalledWith('0', 1, true);
-                expect(GoogleEarthEnterpriseMetadata.prototype.getQuadTreePacket).toHaveBeenCalledWith('01', 1, true);
-                expect(GoogleEarthEnterpriseMetadata.prototype.getQuadTreePacket).toHaveBeenCalledWith('012', 1, true);
+                expect(GoogleEarthEnterpriseMetadata.prototype.getQuadTreePacket).toHaveBeenCalledWith('', 1);
+                expect(GoogleEarthEnterpriseMetadata.prototype.getQuadTreePacket).toHaveBeenCalledWith('0', 1, request);
+                expect(GoogleEarthEnterpriseMetadata.prototype.getQuadTreePacket).toHaveBeenCalledWith('01', 1, request);
+                expect(GoogleEarthEnterpriseMetadata.prototype.getQuadTreePacket).toHaveBeenCalledWith('012', 1, request);
 
                 var tileInfo = metadata._tileInfo;
                 expect(tileInfo['0']).toBeDefined();